tmsu-dired-query: Accept the query as a list, not a string
Warning!  It's a breaking change for the queries previously saved as
bookmarks!  The old org-mode links should be safe as they need to
be (de)serialized anyway.

The main rationale for this change is for the command history to be
more consistent with the completing-read-multiple format.  Now we
operate on lists everywhere.