Replace shell-command with make-process
shell-command was problematic with some TRAMP backends (sshx was
broken, sudo was working).  The process was starting immediately
instead of allowing the code below to set the filter and sentinel.
make-process allows to set them immediately.

The command itself is still being passed to the shell because I don't
feel like handling piping by hand.